Add files via upload
This file holds the complete details about the timeline of how pixel play dataset was transformed into a star schema. It highlights the process of dimensional modelling and also details the thought process in figuring out our fact measures in the fact tables. (Holds a snapshot of our star schema design)